Rename migration script
