+++ packages/core/src/scripts/migrate-oz-cli-project.ts
