+++ packages/plugin-truffle/package.json
@@ -8,7 +8,7 @@
     "/dist"
   ],
   "bin": {
-    "migrate-oz-sdk-project": "dist/scripts/migrate-oz-sdk-project.js"
+    "migrate-oz-cli-project": "dist/scripts/migrate-oz-cli-project.js"
   },
   "scripts": {
     "prepublish": "rimraf dist *.tsbuildinfo",
