+++ packages/plugin-truffle/src/scripts/migrate-oz-cli-project.ts
